python_scripts/hybrid_health_engine/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2

# Deep Learning and Computer Vision
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.3
pillow>=8.3.1

# API and Async
aiohttp>=3.7.4
requests>=2.26.0
jira>=3.2.0
PyGithub>=1.55

# Data processing and utilities
python-dotenv>=0.19.0
pyyaml>=5.4.1

# Optional: Visualization
matplotlib>=3.4.3
seaborn>=0.11.2